hw6: Move ServiceContainer exceptions to separated namespace
