+++ application/base/Exceptions/ServiceContainer/ServiceContainerException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Base\Exceptions;
+namespace App\Base\Exceptions\ServiceContainer;
 
 use Psr\Container\ContainerExceptionInterface;
 use RuntimeException;
